requirements.txt
opencv-python
numpy
torch
torchvision
transformers
Pillow
streamlit
plotly
requests